+++ packages/react-appkit/src/containers/AppBar.js
@@ -304,9 +304,9 @@ const AppBar = ({
   //   menuItems.push(action(ACTION_IMPORT_KEYRING));
   // }
 
-  // if (onSettingsOpened) {
-  //   menuItems.push(action(ACTION_OPEN_SETTINGS));
-  // }
+  if (onSettingsOpened) {
+    menuItems.push(action(ACTION_OPEN_SETTINGS));
+  }
 
   // if (onPartyHomeNavigation) {
   //   menuItems.push(action(ACTION_OPEN_PARTY_HOME));
@@ -434,7 +434,7 @@ const AppBar = ({
         ))}
 
         <div>
-          <Tooltip title={humanize(profile.username)}>
+          <Tooltip title={profile.username || humanize(profile.publicKey)}>
             <IconButton color='inherit'>
               <ProfileIcon />
             </IconButton>