+++ packages/react-appkit/src/containers/AppKitProvider.tsx
@@ -2,17 +2,15 @@
 // Copyright 2020 DXOS.org
 //
 
-import React, { useEffect, useReducer } from 'react';
+import React, { ReactNode, useEffect, useReducer, useState } from 'react';
 import defaultsDeep from 'lodash.defaultsdeep';
-
-import { useReset, useConfig } from '@dxos/react-client';
-import { ErrorBoundary } from '@dxos/react-ux';
 
 import errorsReducer, { SET_ERRORS } from '../hooks/errors';
 import filterReducer, { SET_FILTER } from '../hooks/filter';
 import layoutReducer, { SET_LAYOUT } from '../hooks/layout';
 
 import { AppKitContext, DefaultRouter } from '../hooks';
+import { useClient } from '@dxos/react-client';
 
 const defaultState = {
   [SET_LAYOUT]: {
@@ -27,48 +25,44 @@ const defaultState = {
  * @param {Object} state
  * @param {string} action
  */
-const appReducer = (state, action) => ({
+const appReducer = (state: any, action: any) => ({
   // TODO(burdon): Key shouldn't be same as action type.
   [SET_ERRORS]: errorsReducer(state[SET_ERRORS], action),
   [SET_FILTER]: filterReducer(state[SET_FILTER], action),
   [SET_LAYOUT]: layoutReducer(state[SET_LAYOUT], action)
 });
 
+export interface AppKitProviderProps {
+  children?: ReactNode,
+  initialState?: {
+    [SET_ERRORS]?: any,
+    [SET_FILTER]?: any,
+    [SET_LAYOUT]?: any,
+  },
+  pads?: any[],
+  errorHandler?: any,
+  issuesLink?: string,
+  router?: any,
+}
+
 /**
  * Creates the AppKit framework context, which provides the global UX state.
  * Wraps children with a React ErrorBoundary component, which catches runtime errors and enables reset.
- *
- * @param {function} children
- * @param {Object} [initialState]
- * @param {Object} [router]
- * @param {function} [errorHandler]
- * @param {Array} [pads]
- * @returns {function}
  */
-const AppKitContextProvider = ({ children, initialState, router = DefaultRouter, errorHandler, pads = [], issuesLink = undefined }) => {
+const AppKitProvider = ({ children, initialState, router = DefaultRouter, errorHandler, pads = [], issuesLink = undefined }: AppKitProviderProps) => {
+  const client = useClient();
+
   const [state, dispatch] = useReducer(appReducer, defaultsDeep({}, initialState, defaultState));
-  const reset = useReset();
-  const config = useConfig();
+  const [padsRegistered, setPadsRegistered] = useState(false);
 
   const { errors: { exceptions = [] } = {} } = state[SET_ERRORS] || {};
-
-  const handleRestart = () => {
-    window.location.assign(config.app.publicUrl || '/');
-  };
-
-  const handleReset = async () => {
-    await reset();
-    handleRestart();
-  };
-
-  const handleError = (...args) => console.error(...args);
 
   // Bind the error handler.
   useEffect(() => {
     if (!errorHandler) {
       return;
     }
-    errorHandler.on('error', error => {
+    errorHandler.on('error', (error: any) => {
       dispatch({
         type: SET_ERRORS,
         payload: {
@@ -78,13 +72,21 @@ const AppKitContextProvider = ({ children, initialState, router = DefaultRouter,
     });
   }, []);
 
+  useEffect(() => {
+    const registerPadModels = async () => {
+      for (const pad of pads) {
+        await pad.register?.(client);
+      }
+      setPadsRegistered(true);
+    };
+    registerPadModels();
+  }, []);
+ 
   return (
     <AppKitContext.Provider value={{ state, dispatch, router, pads, issuesLink }}>
-      <ErrorBoundary onError={handleError} onRestart={handleRestart} onReset={handleReset}>
-        {children}
-      </ErrorBoundary>
+      {padsRegistered && children}
     </AppKitContext.Provider>
   );
 };
 
-export default AppKitContextProvider;
+export default AppKitProvider;