+++ packages/react-appkit/src/containers/PartyCardContainer.js
@@ -5,38 +5,27 @@
 import React, { useState } from 'react';
 
 import { keyToBuffer, keyToString } from '@dxos/crypto';
-import { ObjectModel } from '@dxos/object-model';
 import { useClient, useItems } from '@dxos/react-client';
 
 import PartyCard from '../components/PartyCard';
 import { useAppRouter, usePads } from '../hooks';
 import DefaultSettingsDialog from './DefaultSettingsDialog';
 
-const PartyCardContainer = ({ party, ipfs }) => {
+const PartyCardContainer = ({ party }) => {
   const client = useClient();
   const router = useAppRouter();
   const [pads] = usePads();
   const topic = keyToString(party.key);
-  // const { model, createItem } = useItems(topic);
   const items = useItems({ partyKey: keyToBuffer(topic), type: pads.map(pad => pad.type) });
   const [newItemType, setNewItemType] = useState(undefined);
   const [itemSettingsOpen, setItemSettingsOpen] = useState(false);
 
   const handleSavedSettings = async ({ name }, metadata = {}, callback) => {
     const pad = pads.find(p => p.type === newItemType);
-    if (pad && pad.create) {
-      const itemId = await pad.create({ party, client }, { name });
-      router.push({ topic, item: itemId });
-    } else {
-      const item = await party.database.createItem({
-        model: ObjectModel,
-        type: newItemType,
-        props: { title: name || 'random-name' }
-      });
-      callback && callback(item.id);
-      router.push({ topic, item: item.id });
-    }
+    const item = await pad.create({ party, client }, { name }, metadata);
+    callback && callback(item);
     handleCanceledSettings();
+    router.push({ topic, item: item.id });
   };
 
   const handleCanceledSettings = () => {
@@ -57,7 +46,7 @@ const PartyCardContainer = ({ party, ipfs }) => {
       <PartyCard
         client={client}
         party={party}
-        itemModel={items}
+        items={items}
         router={router}
         pads={pads}
         onNewItemRequested={handleNewItemRequested}
@@ -70,7 +59,6 @@ const PartyCardContainer = ({ party, ipfs }) => {
         onClose={handleSavedSettings}
         onCancel={handleCanceledSettings}
         item={undefined} // no item!
-        // itemModel={model}
       />
     </>
   );