+++ packages/react-appkit/stories/decorators/WithClientProvider.js
@@ -2,53 +2,36 @@
 // Copyright 2020 DXOS.org
 //
 
-import React, { useState, useEffect } from 'react';
-
-import { Client } from '@dxos/client';
-import { ClientProvider } from '@dxos/react-client';
-import { createStorage } from '@dxos/random-access-multi-storage';
-import { Keyring, KeyType } from '@dxos/credentials';
-import { createSchema, Registry, DEFAULT_CHAIN_ID } from '@wirelineio/registry-client';
-import { config, registryData } from '../common';
-
-const storage = createStorage('./db/stories', 'ram');
-
-export const WithClient = (story) => {
-  const client = new Client({ storage });
-  return (
-    <RenderProvider story={story} client={client} config={config} />
-  );
-};
+import { createKeyPair } from '@dxos/crypto';
+import { useClient } from '@dxos/react-client';
+import React, { useEffect, useState } from 'react';
+import { ClientInitializer } from '../../src/containers/ClientInitializer';
+import { config } from '../common';
 
 export const WithClientAndIdentity = (story) => {
-  const [client, setClient] = useState(false);
-
-  useEffect(() => {
-    async function runEffect () {
-      const keyring = new Keyring();
-      await keyring.createKeyRecord({ type: KeyType.IDENTITY });
-      const schema = await createSchema(registryData);
-      const registry = new Registry(undefined, DEFAULT_CHAIN_ID, { schema });
-      // TODO(rzadp,rburdon): Replace with actual client SDK for creating a profile
-      const client = new Client({ storage, keyring, registry });
-      await client.initialize();
-      setClient(client);
-    }
-    runEffect();
-  }, []);
   return (
-    <>
-      {client && (
-        <RenderProvider story={story} client={client} config={config} />
-      )}
-    </>
+    <ClientInitializer config={config}>
+      <RenderProvider story={story} />
+    </ClientInitializer>
   );
 };
 
-function RenderProvider ({ story, ...props }) {
+function RenderProvider ({ story }) {
+  const [ready, setReady] = useState(false);
+  const client = useClient();
+
+  useEffect(() => {
+    (async () => {
+      await client.createProfile({ ...createKeyPair(), username: 'foo' });
+      setReady(true);
+    })();
+  }, []);
+
+  if (!ready) {
+    return null;
+  }
+
   return (
-    <ClientProvider {...props}>
-      <div className='WithClientDecorator'>{story()}</div>
-    </ClientProvider>
+    <div className='WithClientDecorator'>{story()}</div>
   );
 }