+++ CHANGELOG.md
@@ -1,5 +1,116 @@
 # Changelog
 
+### [2.6.5](https://www.github.com/dxos/sdk/compare/v2.6.4...v2.6.5) (2020-10-28)
+
+
+### Bug Fixes
+
+* Fix react state leak ([#269](https://www.github.com/dxos/sdk/issues/269)) ([15a4e80](https://www.github.com/dxos/sdk/commit/15a4e80d3cf9b8427b6a36780760bc7d415b700d))
+* Replace error view ([#266](https://www.github.com/dxos/sdk/issues/266)) ([d2f1b83](https://www.github.com/dxos/sdk/commit/d2f1b83579b07bfcd9596b4796c9b5c8a8bb6c29))
+
+### [2.6.4](https://www.github.com/dxos/sdk/compare/v2.6.3...v2.6.4) (2020-10-26)
+
+
+### Bug Fixes
+
+* Fix broken export ([7867dca](https://www.github.com/dxos/sdk/commit/7867dca9a1e16fe978ce5092cff0239c681e07a8))
+
+### [2.6.3](https://www.github.com/dxos/sdk/compare/v2.6.2...v2.6.3) (2020-10-26)
+
+
+### Bug Fixes
+
+* Fix forgotten export ([ca843ff](https://www.github.com/dxos/sdk/commit/ca843ff645f3a5395a68684b3cb4e2a2a19ee763))
+
+### [2.6.2](https://www.github.com/dxos/sdk/compare/v2.6.1...v2.6.2) (2020-10-26)
+
+
+### Bug Fixes
+
+* Transpile nullish coalescing and optional chaining ([e1456a2](https://www.github.com/dxos/sdk/commit/e1456a2c0cb530bfd7849ac4b1810dc0db8d8f3c))
+
+### [2.6.1](https://www.github.com/dxos/sdk/compare/v2.6.0...v2.6.1) (2020-10-26)
+
+
+### Bug Fixes
+
+* Fix 'main' & 'types' fileds in appkit ([218e92a](https://www.github.com/dxos/sdk/commit/218e92a44426badf70749469351b65aa25c82c1c))
+
+## [2.6.0](https://www.github.com/dxos/sdk/compare/v2.5.2...v2.6.0) (2020-10-26)
+
+
+### Features
+
+* Add pad registration to AppKitProvider ([16de2dc](https://www.github.com/dxos/sdk/commit/16de2dc489c760739eaec79be196a610ef4e8c2f))
+* Rename AppKitContextProvider -> AppKitProvider ([3879765](https://www.github.com/dxos/sdk/commit/38797653686e57702912c24ee60acf294f804e50))
+
+
+### Bug Fixes
+
+* Add 'persistent' storage type that selects the default driver for the platform ([be3fc2d](https://www.github.com/dxos/sdk/commit/be3fc2d7910d221a5782a71d5a35772cc91d24e0))
+* Deprecate check for errors ([35e0a40](https://www.github.com/dxos/sdk/commit/35e0a40140b1b72792694940c1effa43b9a11e38))
+
+### [2.5.2](https://www.github.com/dxos/sdk/compare/v2.5.1...v2.5.2) (2020-10-26)
+
+
+### Features
+
+* Rewrite client config ([#258](https://www.github.com/dxos/sdk/issues/258)) ([d52d8e9](https://www.github.com/dxos/sdk/commit/d52d8e9f0937d37d4d9c4e586c48e8966ba60fd5))
+
+
+### Bug Fixes
+
+* Preserve done invitation state ([#256](https://www.github.com/dxos/sdk/issues/256)) ([945fe0f](https://www.github.com/dxos/sdk/commit/945fe0fb7b619cb2a2dcd6813ec0c9afe552aef2))
+
+### [2.5.1](https://www.github.com/dxos/sdk/compare/v2.5.0...v2.5.1) (2020-10-23)
+
+
+### Features
+
+* Restore party item restore functionality ([#253](https://www.github.com/dxos/sdk/issues/253)) ([4ea6819](https://www.github.com/dxos/sdk/commit/4ea6819c51ed723f3e4bd63a24e8871f4cb29583))
+
+
+### Bug Fixes
+
+* automatically close redeem dialog and restore state to initial ([#255](https://www.github.com/dxos/sdk/issues/255)) ([662a493](https://www.github.com/dxos/sdk/commit/662a493f63365310389cc2541d5db06bf87a6836))
+* Increate client initialization warning timeout  ([#252](https://www.github.com/dxos/sdk/issues/252)) ([1920095](https://www.github.com/dxos/sdk/commit/1920095f6675a76d916baa81269ca9f974c35f4a))
+
+## [2.5.0](https://www.github.com/dxos/sdk/compare/v2.4.0...v2.5.0) (2020-10-22)
+
+
+### Features
+
+* Bring back generic pad settings ([#231](https://www.github.com/dxos/sdk/issues/231)) ([0ea5a36](https://www.github.com/dxos/sdk/commit/0ea5a3694e969563c1de0c195a3ab7dab2f3a522))
+
+
+### Bug Fixes
+
+* Default party display name ([#247](https://www.github.com/dxos/sdk/issues/247)) ([2dac75a](https://www.github.com/dxos/sdk/commit/2dac75a85c7721f07324b67b0a71f99b3bee3783))
+* Disable broken party settings ([#246](https://www.github.com/dxos/sdk/issues/246)) ([8c7874a](https://www.github.com/dxos/sdk/commit/8c7874a763c5593be9a5291c43fd924e18dac381))
+* Fix effects firing on every render ([239c720](https://www.github.com/dxos/sdk/commit/239c7200a55facaceddc2f731a85d8c02cb91b70))
+* Pass pad metadata to pad creation function. Remove generic pad creation function. ([#234](https://www.github.com/dxos/sdk/issues/234)) ([25d905d](https://www.github.com/dxos/sdk/commit/25d905d44f2bd0480da46c8add3c3b052e581a80))
+* Set devtools hook in client provider & add persistent storage story ([#237](https://www.github.com/dxos/sdk/issues/237)) ([53a1180](https://www.github.com/dxos/sdk/commit/53a11800868face50451d795b0a89649be40d272))
+* update botkit deps. ([#249](https://www.github.com/dxos/sdk/issues/249)) ([cff72c6](https://www.github.com/dxos/sdk/commit/cff72c62ce0d39210e2f39a649dfa08edfccb2b1))
+
+## [2.4.0](https://www.github.com/dxos/sdk/compare/v2.3.1...v2.4.0) (2020-10-21)
+
+
+### Features
+
+* react-client refactoring & stability improvements ([#227](https://www.github.com/dxos/sdk/issues/227)) ([9bf0e67](https://www.github.com/dxos/sdk/commit/9bf0e6789b107eb1f0065cfe8219243fd0755637))
+
+
+### Bug Fixes
+
+* Fix queue microtask shim ([36f0b8d](https://www.github.com/dxos/sdk/commit/36f0b8d436dd194d3d0858774ed9cbf34b9e3cfa))
+
+### [2.3.1](https://www.github.com/dxos/sdk/compare/v2.3.0...v2.3.1) (2020-10-20)
+
+
+### Bug Fixes
+
+* bot to emit events on party join. ([#222](https://www.github.com/dxos/sdk/issues/222)) ([02fde00](https://www.github.com/dxos/sdk/commit/02fde00612e2068849cdb3a57b16d051793b5235))
+
 ## [2.3.0](https://www.github.com/dxos/sdk/compare/v2.2.1...v2.3.0) (2020-10-19)